Rename module: server -> mainWindow
